Fixing misused BPA options and fixing header declarations
