+++ mock-bpa-test/test_bpa.py
@@ -79,7 +79,7 @@ class TestAgent(unittest.TestCase):
 
         args = compose_args([
             'bsl-mock-bpa',
-            '-e', 'ipn:2.1',
+            '-s', 'ipn:2.1',  # security source
             '-u', 'localhost:4556', '-r', 'localhost:14556',
             '-o', 'localhost:24556', '-a', 'localhost:34556',
             '-p', policy_config,
