+++ src/mock_bpa/eid.c
@@ -33,13 +33,6 @@
 #include <stdio.h>
 #include <sys/types.h>
 
-int MockBPA_GetEid(const void *user_data, BSL_HostEID_t *result_eid)
-{
-    const char *local_ipn = getenv("BSL_TEST_LOCAL_IPN_EID");
-    int         x         = mock_bpa_eid_from_text(result_eid, local_ipn, (void *)user_data);
-    return (0 == x) ? 0 : -1;
-}
-
 void bsl_mock_eid_init(bsl_mock_eid_t *eid)
 {
     CHKVOID(eid);
